fix: перенес событие по esc в modal
