+++ src/components/modal/ModalOverlay.tsx
@@ -1,4 +1,3 @@
-import { useEffect } from 'react';
 import cl from './ModalOverlay.module.css';
 
 export default function ModalOverlay({
@@ -8,19 +7,6 @@ export default function ModalOverlay({
 	children: React.ReactNode;
 	onClose: () => void;
 }) {
-	useEffect(() => {
-		const handleEscDown = (event: KeyboardEvent) => {
-			if (event.key === 'Escape') {
-				onClose();
-			}
-		};
-
-		document.addEventListener('keydown', handleEscDown);
-
-		return () => {
-			document.removeEventListener('keydown', handleEscDown);
-		};
-	}, []);
 	return (
 		<div className={cl.modalOverlay} onClick={onClose}>
 			<div className={cl.modalContent} onClick={(e) => e.stopPropagation()}>
